LayoutTests/fast/webgpu/ftoi.html

<script src="../../resources/js-test-pre.js"></script>
<script>

var output;
const BUFFER_SIZE = 64 * 4;

const SHADER = `
@group(0) @binding(0) var<storage, read_write> buf: array<atomic<u32>, 64>;

@compute @workgroup_size(1)
fn main() {
  var f = -1f;
  atomicStore(&buf[u32(f)], 0x1234);
}
`;


async function main() {
    const adapter = await navigator.gpu.requestAdapter({});
    const device = await adapter.requestDevice({});

    device.pushErrorScope('validation');

    const shaderModule = device.createShaderModule({code: SHADER});
    const bindGroupLayout = device.createBindGroupLayout({
        entries: [
            {binding: 0, buffer: {type: 'storage'}, visibility: GPUShaderStage.COMPUTE},
        ]
    });

    const oobBuffer = device.createBuffer({
        usage: GPUBufferUsage.STORAGE | GPUBufferUsage.COPY_SRC,
        size: BUFFER_SIZE
    });

    const outputBuffer = device.createBuffer({
        usage: GPUBufferUsage.COPY_DST | GPUBufferUsage.MAP_READ,
        size: BUFFER_SIZE
    });

    const bindGroup = device.createBindGroup({
        layout: bindGroupLayout, entries: [
            {binding: 0, resource: {buffer: oobBuffer}},
        ],
    });

    const commandEncoder = device.createCommandEncoder();
    const computePassEncoder = commandEncoder.beginComputePass({});
    const pipelineLayout = device.createPipelineLayout({bindGroupLayouts: [bindGroupLayout]});
    const computePipeline = device.createComputePipeline({layout: pipelineLayout, compute: {module: shaderModule}});
    computePassEncoder.setPipeline(computePipeline);
    computePassEncoder.setBindGroup(0, bindGroup);
    computePassEncoder.dispatchWorkgroups(1);
    computePassEncoder.end();

    commandEncoder.copyBufferToBuffer(oobBuffer, 0, outputBuffer, 0, BUFFER_SIZE);

    device.queue.submit([commandEncoder.finish()]);

    await device.queue.onSubmittedWorkDone();

    await outputBuffer.mapAsync(GPUMapMode.READ);
    output = new Uint32Array(outputBuffer.getMappedRange());

    debug(output);

    shouldBe('output[0]', '0x1234');
    shouldBe('output[63]', '0');

    outputBuffer.unmap();
}


globalThis.testRunner?.waitUntilDone();

main().catch(e => {
    debug(e);
}).finally(() => {
    globalThis.testRunner?.notifyDone();
});

</script>
